requirements.txt
azure-ai-projects>=2.0.0b1
azure-identity>=1.19.0
openai>=1.0.0
pyyaml>=6.0
